+++ server/UserData/Item/item.cs
@@ -12,11 +12,11 @@ namespace Nanina.UserData.ItemData
     public class Item
     {
         public ushort id {get; set;}
-        public uint inventoryId {get; set;}
-        public uint count {get; set;}
+        public int inventoryId {get; set;}
+        public int count {get; set;}
         public ItemType type {get; set;}
         public required string imgPATH {get; set;}
-        public byte rarity {get; set;}
+        public short rarity {get; set;}
         public required List<Modifier> modifiers {get; set;}
     }
 }
